requirements.txt
transformers==4.36.2
torch==2.1.2
python-telegram-bot==20.7